look for scheduled reboot instead of pending
I misunderstood what the built-in `reboot_pending?` function does. It looks to determine
if a host needs rebooted after a package update has been installed and does not strictly
mean that a reboot is scheduled via `shutdown`, which is what I really wanted to test for.

dbus systems support fetching whether a reboot is scheduled. I've effectively only whitelisted
this for Ubuntu systems because that's all I really presently care about but should be able
to be extended to other systems easily if the desire is there.